pyproject.toml
[tool.poetry]
name = "ytscrap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Scrapy = "^2.4.1"
dataset = "^1.4.1"
PyYAML = "^5.3.1"
munch = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.4"
autoflake = "^1.4"
isort = "^5.6.4"
ipython = "^7.19.0"

[tool.poetry.scripts]
vtube= 'ytscrap.client:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
